feat: divide controller
