+++ application/server/src/main/kotlin/io/github/siyual_park/application/server/dto/response/UserContactInfo.kt
@@ -1,0 +1,9 @@
+package io.github.siyual_park.application.server.dto.response
+
+import java.time.Instant
+
+data class UserContactInfo(
+    val email: String,
+    val createdAt: Instant,
+    val updatedAt: Instant?
+)
