+++ application/server/src/test/kotlin/io/github/siyual_park/application/server/controller/UserControllerTest.kt
@@ -2,13 +2,11 @@ package io.github.siyual_park.application.server.controller
 
 import io.github.siyual_park.IntegrationTest
 import io.github.siyual_park.application.server.dto.request.GrantScopeRequest
-import io.github.siyual_park.application.server.dto.request.UpdateUserCredentialRequest
 import io.github.siyual_park.application.server.dto.request.UpdateUserRequest
 import io.github.siyual_park.application.server.dummy.DummyCreateClientPayload
 import io.github.siyual_park.application.server.dummy.DummyCreateUserPayload
 import io.github.siyual_park.application.server.dummy.DummyCreateUserRequest
 import io.github.siyual_park.application.server.dummy.DummyNameFactory
-import io.github.siyual_park.application.server.dummy.DummyStringFactory
 import io.github.siyual_park.application.server.gateway.GatewayAuthorization
 import io.github.siyual_park.application.server.gateway.UserControllerGateway
 import io.github.siyual_park.auth.domain.scope_token.ScopeTokenFactory
@@ -676,94 +674,6 @@ class UserControllerTest @Autowired constructor(
         )
 
         val response = userControllerGateway.revokeScope(otherUser.id, scope.id)
-
-        assertEquals(HttpStatus.FORBIDDEN, response.status)
-    }
-
-    @Test
-    fun `PATCH users_{self-id}_credential, status = 200`() = blocking {
-        val payload = DummyCreateUserPayload.create()
-        val user = userFactory.create(payload)
-        val principal = user.toPrincipal()
-
-        gatewayAuthorization.setPrincipal(
-            principal,
-            push = listOf("users[self].credential:update")
-        )
-
-        val password = DummyStringFactory.create(10)
-        val request = UpdateUserCredentialRequest(
-            password = Optional.of(password)
-        )
-        val response = userControllerGateway.updateCredential(user.id, request)
-
-        assertEquals(HttpStatus.OK, response.status)
-
-        response.responseBody.awaitSingle()
-    }
-
-    @Test
-    fun `PATCH users_{self-id}_credential, status = 403`() = blocking {
-        val payload = DummyCreateUserPayload.create()
-        val user = userFactory.create(payload)
-        val principal = user.toPrincipal()
-
-        gatewayAuthorization.setPrincipal(
-            principal,
-            pop = listOf("users[self].credential:update", "users.credential:update")
-        )
-        val password = DummyStringFactory.create(10)
-        val request = UpdateUserCredentialRequest(
-            password = Optional.of(password)
-        )
-        val response = userControllerGateway.updateCredential(user.id, request)
-
-        assertEquals(HttpStatus.FORBIDDEN, response.status)
-    }
-
-    @Test
-    fun `PATCH users_{user-id}_credential, status = 200`() = blocking {
-        val payload = DummyCreateUserPayload.create()
-        val user = userFactory.create(payload)
-        val principal = user.toPrincipal()
-
-        val otherUser = DummyCreateUserPayload.create()
-            .let { userFactory.create(it) }
-
-        gatewayAuthorization.setPrincipal(
-            principal,
-            push = listOf("users.credential:update")
-        )
-
-        val password = DummyStringFactory.create(10)
-        val request = UpdateUserCredentialRequest(
-            password = Optional.of(password)
-        )
-        val response = userControllerGateway.updateCredential(otherUser.id, request)
-
-        assertEquals(HttpStatus.OK, response.status)
-
-        response.responseBody.awaitSingle()
-    }
-
-    @Test
-    fun `PATCH users_{user-id}_credential, status = 403`() = blocking {
-        val payload = DummyCreateUserPayload.create()
-        val user = userFactory.create(payload)
-        val principal = user.toPrincipal()
-
-        val otherUser = DummyCreateUserPayload.create()
-            .let { userFactory.create(it) }
-
-        gatewayAuthorization.setPrincipal(
-            principal,
-            pop = listOf("users[self].credential:update", "users.credential:update")
-        )
-        val password = DummyStringFactory.create(10)
-        val request = UpdateUserCredentialRequest(
-            password = Optional.of(password)
-        )
-        val response = userControllerGateway.updateCredential(otherUser.id, request)
 
         assertEquals(HttpStatus.FORBIDDEN, response.status)
     }
