+++ application/server/src/test/kotlin/io/github/siyual_park/application/server/gateway/UserControllerGateway.kt
@@ -2,10 +2,8 @@ package io.github.siyual_park.application.server.gateway
 
 import io.github.siyual_park.application.server.dto.request.CreateUserRequest
 import io.github.siyual_park.application.server.dto.request.GrantScopeRequest
-import io.github.siyual_park.application.server.dto.request.UpdateUserCredentialRequest
 import io.github.siyual_park.application.server.dto.request.UpdateUserRequest
 import io.github.siyual_park.application.server.dto.response.ScopeTokenInfo
-import io.github.siyual_park.application.server.dto.response.UserCredentialInfo
 import io.github.siyual_park.application.server.dto.response.UserInfo
 import io.github.siyual_park.ulid.ULID
 import org.springframework.http.HttpHeaders
@@ -111,14 +109,5 @@ class UserControllerGateway(
             .header(HttpHeaders.AUTHORIZATION, gatewayAuthorization.getAuthorization())
             .exchange()
             .returnResult()
-    }
-
-    suspend fun updateCredential(userId: ULID, request: UpdateUserCredentialRequest): FluxExchangeResult<UserCredentialInfo> {
-        return client.patch()
-            .uri("/users/$userId/credential")
-            .header(HttpHeaders.AUTHORIZATION, gatewayAuthorization.getAuthorization())
-            .bodyValue(request)
-            .exchange()
-            .returnResult(UserCredentialInfo::class.java)
     }
 }
